app/requirements.txt
fastapi>=0.71.0,<0.72.0
uvicorn>=0.16.0,<0.17.0
python-dotenv>=0.19.2,<0.20.0
motor>=2.5.1,<2.6.0
dnspython>=2.1.0,<2.2.0
